requirements.txt
livekit
Pillow
numpy
python-dotenv 